+++ src/main/java/kr/hhplus/be/server/domain/order/OrderItem.java
@@ -1,0 +1,11 @@
+package kr.hhplus.be.server.domain.order;
+
+public record OrderItem(
+    long id,
+    long orderId,
+    long productId,
+    int quantity,
+    int price
+) {
+
+}
