+++ src/main/java/kr/hhplus/be/server/infrastructure/order/OrderRepository.java
@@ -1,0 +1,5 @@
+package kr.hhplus.be.server.infrastructure.order;
+
+public interface OrderRepository {
+
+}
