+++ src/main/java/kr/hhplus/be/server/infrastructure/order/OrderRepositoryImpl.java
@@ -1,0 +1,10 @@
+package kr.hhplus.be.server.infrastructure.order;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class OrderRepositoryImpl implements OrderRepository {
+
+    // TODO OrderJpaRepository 와 같은 CRUD 해올 객체 추가
+
+}
