+++ src/main/java/kr/hhplus/be/server/interfaces/order/OrderApi.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order;
+package kr.hhplus.be.server.interfaces.order;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -7,6 +7,8 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import kr.hhplus.be.server.domain.order.Order;
+import kr.hhplus.be.server.interfaces.order.OrderRequest.CreateOrder;
 import org.springframework.web.ErrorResponse;
 
 @Tag(name = "Order", description = "주문 관련 API")
@@ -20,6 +22,6 @@ public interface OrderApi {
         @ApiResponse(responseCode = "500", description = "서버 에러",
             content = @Content(schema = @Schema(implementation = ErrorResponse.class)))
     })
-    Order createOrder(long userId, Order order);
+    OrderResponse.Order createOrder(long userId, CreateOrder createOrder);
 
 }
